Update revision 08-07-2019

+++ MK4duo/src/lcd/ultralcd/dogm/ultralcd_dogm.cpp
@@ -142,12 +142,12 @@ void LcdUI::set_font(const MK4duoFontEnum font_nr) {
     // Determine text space needed
     #if DISABLED(STRING_SPLASH_LINE2)
       constexpr uint8_t text_total_height = MENU_FONT_HEIGHT,
-                        text_width_1 = (sizeof(STRING_SPLASH_LINE1) - 1) * (MENU_FONT_WIDTH),
+                        text_width_1 = uint8_t(sizeof(STRING_SPLASH_LINE1) - 1) * uint8_t(MENU_FONT_WIDTH),
                         text_width_2 = 0;
     #else
-      constexpr uint8_t text_total_height = (MENU_FONT_HEIGHT) * 2,
-                        text_width_1 = (sizeof(STRING_SPLASH_LINE1) - 1) * (MENU_FONT_WIDTH),
-                        text_width_2 = (sizeof(STRING_SPLASH_LINE2) - 1) * (MENU_FONT_WIDTH);
+      constexpr uint8_t text_total_height = uint8_t(MENU_FONT_HEIGHT) * 2,
+                        text_width_1 = uint8_t(sizeof(STRING_SPLASH_LINE1) - 1) * uint8_t(MENU_FONT_WIDTH),
+                        text_width_2 = uint8_t(sizeof(STRING_SPLASH_LINE2) - 1) * uint8_t(MENU_FONT_WIDTH);
     #endif
     constexpr uint8_t text_max_width = MAX(text_width_1, text_width_2),
                       rspace = width - (START_BMPWIDTH);
